Code refactored. Each functionality(e.g. Filter, Map, PhotoDisplay) has its own folder to store the related files. 'npm run build' runs successfully
