+++ components/Filter/KeywordSelect.tsx
@@ -1,21 +1,18 @@
 /* eslint-disable react/display-name */
-import React, { useState,useRef, useImperativeHandle, forwardRef } from 'react';
 import styles from '../../styles/Home.module.css';
-
 import Checkbox from '@mui/material/Checkbox';
 import TextField from '@mui/material/TextField';
 import Autocomplete from '@mui/material/Autocomplete';
 import CheckBoxOutlineBlankIcon from '@mui/icons-material/CheckBoxOutlineBlank';
 import CheckBoxIcon from '@mui/icons-material/CheckBox';
 
- export default function KeywordsSelect(props: any) {
-      
+export default function KeywordsSelect(props: any) {      
     return (
       <Autocomplete
       className={styles.controls}
         multiple
         id="checkboxes-tags-demo"
-        options={props.keywords}
+        options={Object.keys(props.keywordsObject)}
         disableCloseOnSelect
         onChange={(event, newValue) => {
             props.changeSelectedKeywords(newValue)
