+++ components/Map/Map.tsx
@@ -4,7 +4,7 @@ import 'leaflet/dist/leaflet.css'
 import 'leaflet-defaulticon-compatibility/dist/leaflet-defaulticon-compatibility.webpack.css'; // Re-uses images from ~leaflet package
 import 'leaflet-defaulticon-compatibility';
 import { LatLngTuple } from 'leaflet';
-import  { Photo ,MarkerController } from '../lib/types';
+import  { Photo  } from '../../lib/types';
 import PhotoMarker from './PhotoMarker';
 
 interface MapProps {
