+++ package.json
@@ -19,7 +19,8 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-dual-range-slider": "^1.0.4",
-    "react-leaflet": "^3.2.0"
+    "react-leaflet": "^3.2.0",
+    "sharp": "^0.29.1"
   },
   "devDependencies": {
     "@types/leaflet": "^1.7.4",
